move service, change endpoint name
