+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/config/DatasourceConfiguration.kt
@@ -8,6 +8,7 @@ import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Primary
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 import org.springframework.orm.jpa.JpaTransactionManager
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean
@@ -29,6 +30,7 @@ class LibConfiguration {
   fun mainDataSource(mainDataSourceProperties: DataSourceProperties) = mainDataSourceProperties.initializeDataSourceBuilder().build()
 
   @Bean
+  @Primary
   fun mainEntityManagerFactory(
     mainDataSource: DataSource,
     @Value("#{'\${dpr.lib.modelPackages:}'.split(',')}") modelPackages: List<String>,
