+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/missingReport/MissingReportService.kt
@@ -1,10 +1,8 @@
-package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.service
+package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.missingReport
 
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.MissingReportSubmissionRequest
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.missingReport.MissingReportSubmission
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.missingReport.MissingReportSubmissionsRepository
 
 @Service
 @Transactional(transactionManager = "missingReportTransactionManager")
@@ -12,7 +10,7 @@ class MissingReportService(
   private val missingReportSubmissionsRepository: MissingReportSubmissionsRepository,
 ) {
   fun createMissingReportSubmission(
-    missingReportSubmissionRequest: MissingReportSubmissionRequest,
+      missingReportSubmissionRequest: MissingReportSubmissionRequest,
   ): MissingReportSubmission {
     val (userId, reportId, reportVariantId, reason) = missingReportSubmissionRequest
     val submission = MissingReportSubmission(userId, reportId, reportVariantId, reason)
